refactor: add hexagonal architecture
